fix: make voice transcript/completion persistence session-independent
